Extrai conteudo de _app.js e cria novo component Layout
Para facilitar os testes uma vez que o _app.js é um internal do Next o que dificulta os testes
