+++ jest.config.js
@@ -9,6 +9,7 @@ module.exports = {
     '<rootDir>/pages/**/*.js',
     '<rootDir>/hooks/**/*.js',
     '<rootDir>/store/**/*.js',
+    '<rootDir>/layout/**/*.js',
   ],
   watchPlugins: [
     'jest-watch-typeahead/filename',
