+++ pages/index.js
@@ -3,6 +3,7 @@ import { useFetchProducts } from '../hooks/use-fetch-products';
 import ProductCard from '../components/product-card';
 import Search from '../components/search';
 import { useCartStore } from '../store/cart';
+import Layout from '../layout';
 
 export default function Home() {
   const { products, error } = useFetchProducts();
@@ -45,18 +46,20 @@ export default function Home() {
   };
 
   return (
-    <main data-testid="product-list" className="my-8">
-      <Search doSearch={(term) => setTerm(term)} />
-      <div className="container mx-auto px-6">
-        <h3 className="text-gray-700 text-2xl font-medium">Wrist Watch</h3>
-        <span className="mt-3 text-sm text-gray-500">
-          {renderProductQuantity()}
-        </span>
-        <div className="grid gap-6 grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 mt-6">
-          {renderErrorMessage()}
-          {renderProductListOrMessage()}
+    <Layout>
+      <main data-testid="product-list" className="my-8">
+        <Search doSearch={(term) => setTerm(term)} />
+        <div className="container mx-auto px-6">
+          <h3 className="text-gray-700 text-2xl font-medium">Wrist Watch</h3>
+          <span className="mt-3 text-sm text-gray-500">
+            {renderProductQuantity()}
+          </span>
+          <div className="grid gap-6 grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 mt-6">
+            {renderErrorMessage()}
+            {renderProductListOrMessage()}
+          </div>
         </div>
-      </div>
-    </main>
+      </main>
+    </Layout>
   );
 }
